requirements.txt
fastapi
uvicorn
uvicorn
numpy
mysql-connector-python
joblib
onnxruntime
pydantic
python-multipart
requests
